move countLastUpdate to component
