+++ src/components/AddToFavoriteButton/AddToFavoriteButton.jsx
@@ -6,34 +6,29 @@ import { toFavoriteList } from '../../store/chuckApi/types';
 const buttonType = {
     inFavorite: {
         type: 'inFavorite',
-        url: './img/icon/heart.svg',
-        class: 'inFavorite'
+        url: './img/icon/heart.svg'
     },
     notInFavorite: {
         type: 'notInFavorite',
-        url: './img/icon/heart_stroke.svg',
-        class: 'notInFavorite'
+        url: './img/icon/heart_stroke.svg'
     }
 }
 const AddToFavoriteButton = ({ categories, icon_url, id, updated_at, url, inFavorite = undefined, value, favoriteBlockStyle }) => {
     const dispatch = useDispatch();
     const [type, setType] = useState()
     useEffect(() => {
-        if (inFavorite === undefined) {
-            return
-        } else {
-            setType(inFavorite)
-        }
+        return inFavorite !== undefined ? setType(inFavorite) : null
     }, [setType, inFavorite])
 
-    type === true ? <button type='button' className={!favoriteBlockStyle ? style.addIconBtn : style.addIconFavoriteBtn}>
-        <img src={buttonType.inFavorite.url} alt='In favorite button'
-            className={style.inFavorite}
-            onClick={() => {
-                return (dispatch(toFavoriteList(categories, icon_url, id, updated_at, url, value, true)),
-                    setType(false))
-            }} />
-    </button>
+    return type === true ?
+        <button type='button' className={!favoriteBlockStyle ? style.addIconBtn : style.addIconFavoriteBtn}>
+            <img src={buttonType.inFavorite.url} alt='In favorite button'
+                className={style.inFavorite}
+                onClick={() => {
+                    return (dispatch(toFavoriteList(categories, icon_url, id, updated_at, url, value, true)),
+                        setType(false))
+                }} />
+        </button>
         : <button type='button' className={!favoriteBlockStyle ? style.addIconBtn : style.addIconFavoriteBtn}>
             <img src={buttonType.notInFavorite.url} alt="Add to Favorite list"
                 className={style.notInFavorite} onClick={() => {
